Removed canvas-network file
